Shorten server stoned time resulting in pong timeout & refine ircparty support
The assumption is that internet connections are fast to the point that 15min is definitely stoned.

Numerous modifications to ircparty include nicer nicks for users and some nicklist fixes.

Changing ircparty no-nick flag from +z to +s allows +z to be used for its orig purpose in future.

Eventually, if Tcldrop can be made to allow ircparty user nick == botnet nick, it could
moonlight as a somewhat competent IRC bouncer (and too scriptable, at that!)